Reorganised code layout
All purely LB stuff in a separate dir/module.
Separated tests for different parts of LB stuff.
