+++ src/lb/collision.jl
@@ -1,0 +1,8 @@
+type SingleRelaxationTime{T <: FloatingPoint} <: Collision
+    # Inverse of the relaxation time
+    τ⁻¹::T
+end
+
+# Defines collision kernels for SRT
+collision{T}(τ⁻¹::T, fᵢ::Vector{T}, feq::Vector{T}) = τ⁻¹ * (feq - fᵢ)
+collision{T}(k::SingleRelaxationTime{T}, fᵢ::Vector{T}, feq::Vector{T}) = collision(k.τ⁻¹, feq, fᵢ)
