+++ test/lb/collision.jl
@@ -1,0 +1,14 @@
+using LatBo.LatticeBoltzmann: collision
+
+facts("Collision kernels") do
+    context("Single relaxation time") do
+        # random number we can play with and avoid overflow
+        τ⁻¹ = convert(Int64, 2)
+        f = convert(Array{Int64}, rand(Int8, 3))
+        @fact collision(τ⁻¹, f, f) => zeros(Int16, length(f))
+        @fact collision(τ⁻¹*τ⁻¹, f, f) => zeros(Int16, length(f))
+        @fact collision(τ⁻¹, 2*f, f) => -τ⁻¹ * f
+        @fact collision(τ⁻¹, f, 2*f) => τ⁻¹ * f
+        @fact collision(τ⁻¹, 4*f, 2*f) => collision(2*τ⁻¹, 2*f, f)
+    end
+end
