+++ test/lb/lattice.jl
@@ -1,5 +1,4 @@
-using FactCheck: facts, @fact
-using LatBo: D2Q9, D3Q19
+using LatBo.LatticeBoltzmann: D2Q9, D3Q19
 
 facts("Lattice direction inversion") do
     for lattice_name in [:D2Q9, :D3Q19]
